backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain-google-genai
langgraph
tavily-python
python-dotenv